refactor: 각 Repository 구현체마다 인터페이스를 구현하도록 변경
